requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.8
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.2.1
billiard==4.2.1
celery==5.4.0
cerebrus==0.2
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
dill==0.3.9
Django==4.2.19
django-debug-toolbar==5.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
Faker==36.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
isort==6.0.0
kombu==5.4.2
mccabe==0.7.0
msgpack==1.1.0
mysqlclient==2.2.7
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
